Change int to long

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        SalesManager salesManager = new SalesManager(new int[]{10, 20, 54});
+        SalesManager salesManager = new SalesManager(new long[]{10, 20, 54});
         System.out.println("Максимальное значение: " + salesManager.max());
     }
 }
